+++ windows/video_outlet.cc
@@ -29,8 +29,8 @@ VideoOutlet::VideoOutlet(flutter::TextureRegistrar* texture_registrar)
   texture_id_ = texture_registrar_->RegisterTexture(texture_.get());
 }
 
-void VideoOutlet::MarkVideoFrameAvailable(uint8_t* buffer, int32_t width,
-                                          int32_t height) {
+void VideoOutlet::MarkVideoFrameAvailable(uint8_t* buffer, size_t width,
+                                          size_t height) {
   const std::lock_guard<std::mutex> lock(mutex_);
   flutter_pixel_buffer_.buffer = buffer;
   flutter_pixel_buffer_.width = width;
